requirements.txt
Django>=2.0,<2.1
wagtail>=2.1,<2.2
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
django-heroku==0.3.1
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.7
l18n==2016.6.4
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
Unidecode==0.4.21
urllib3==1.23
wagtailfontawesome==1.1.3
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1